@@ -4,7 +4,13 @@
   "description": "Waka GTFS Importer",
   "main": "index.js",
   "scripts": {
-    "start": "node index.js"
+    "type-check": "tsc --noEmit",
+    "type-check:watch": "npm run type-check -- --watch",
+    "build": "npm run build:types && npm run build:js",
+    "build:types": "tsc --emitDeclarationOnly",
+    "build:js": "babel server --out-dir lib --extensions \".ts,.tsx\" --copy-files",
+    "start": "node lib",
+    "bs": "npm run build:js && npm start"
   },
   "repository": {
     "type": "git",
@@ -26,29 +32,39 @@
   },
   "homepage": "https://github.com/dymajo/waka-importer#readme",
   "dependencies": {
-    "aws-sdk": "^2.313.0",
+    "@types/dotenv": "^6.1.1",
+    "@types/extract-zip": "^1.6.2",
+    "@types/mssql": "^4.0.13",
+    "@types/node": "^12.0.0",
+    "@types/rimraf": "^2.0.2",
+    "aws-sdk": "^2.450.0",
+    "axios": "^0.18.0",
     "azure-storage": "^2.10.1",
-    "colors": "^1.3.1",
     "csv-parse": "^2.5.0",
-    "dotenv": "^6.0.0",
+    "dotenv": "^8.0.0",
     "extract-zip": "^1.6.7",
-    "mssql": "^4.1.0",
-    "node-fetch": "^2.3.0",
-    "request": "^2.88.0",
-    "rimraf": "^2.6.2",
-    "stream-transform": "^1.0.2"
+    "mssql": "^5.1.0",
+    "rimraf": "^2.6.3",
+    "stream-transform": "^2.0.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.3.4",
-    "@babel/parser": "^7.3.4",
+    "@babel/cli": "^7.4.4",
+    "@babel/core": "^7.4.4",
+    "@babel/plugin-proposal-class-properties": "^7.4.4",
+    "@babel/plugin-proposal-object-rest-spread": "^7.4.4",
+    "@babel/preset-env": "^7.4.4",
+    "@babel/preset-typescript": "^7.3.3",
+    "@typescript-eslint/eslint-plugin": "^1.7.0",
+    "@typescript-eslint/parser": "^1.7.0",
     "cross-env": "^5.2.0",
-    "eslint": "^5.6.1",
+    "eslint": "^5.16.0",
     "eslint-config": "^0.3.0",
     "eslint-config-airbnb-base": "^13.1.0",
-    "eslint-config-prettier": "^3.1.0",
-    "eslint-plugin-import": "^2.14.0",
-    "eslint-plugin-prettier": "^3.0.0",
-    "eslint-plugin-promise": "^4.0.1",
-    "prettier": "^1.16.4"
+    "eslint-config-prettier": "^4.2.0",
+    "eslint-plugin-import": "^2.17.2",
+    "eslint-plugin-prettier": "^3.0.1",
+    "eslint-plugin-promise": "^4.1.1",
+    "prettier": "^1.17.0",
+    "typescript": "^3.4.5"
   }
 }
