@@ -1,32 +1,31 @@
-const sql = require('mssql')
-const colors = require('colors')
-const config = require('../config')
-
+import { ConnectionPool } from 'mssql'
+import config from '../config'
+import log from '../logger'
 const connectMaster = async () => {
+  const masterConfig = {
+    ...config.db,
+    database: config.db.master_database,
+  }
   const { database } = config.db
-  const masterConfig = JSON.parse(JSON.stringify(config.db))
-  masterConfig.database = masterConfig.master_database
   try {
-    const pool = await sql.connect(masterConfig)
+    const pool = new ConnectionPool(masterConfig)
+    await pool.connect()
     // prepared statements were not working.
     // also, you set this yourself, so your own fault if you drop all your tables
     await pool
       .request()
       .query(`If(db_id(N'${database}') IS NULL) CREATE DATABASE "${database}"`)
   } catch (err) {
-    console.error(
-      'master'.red,
-      'Failed to connect to master database! Check the db.database'.red
-    )
-    console.error(err)
+    log('master', 'Failed to connect to master database! Check the db.database')
+    log(err)
     process.exit(1)
   }
   return true
 }
 
-let cresolve
-let creject
-let pool1
+let cresolve: any
+let creject: any
+let pool1: ConnectionPool
 const ready = new Promise((resolve, reject) => {
   cresolve = resolve
   creject = reject
@@ -36,9 +35,9 @@ const connection = {
   open: () => {
     connectMaster()
       .then(() => {
-        pool1 = new sql.ConnectionPool(config.db, err => {
+        pool1 = new ConnectionPool(config.db, err => {
           if (err) {
-            console.error(err)
+            log(err)
             return creject()
           }
           return cresolve()
@@ -52,4 +51,4 @@ const connection = {
   },
   isReady: ready,
 }
-module.exports = connection
+export default connection
